refactored controllers and fixed resulting problems with specs

+++ app/controllers/questions_controller.rb
@@ -1,10 +1,14 @@
 class QuestionsController < ApplicationController
+  def find_question
+    Question.find(params[:id])
+  end
+
   def index
     @questions = Question.all
   end
 
   def show
-    @question = Question.find(params[:id])
+    @question = find_question
   end
 
   def new
@@ -12,18 +16,14 @@ class QuestionsController < ApplicationController
   end
 
   def edit
-    @question = Question.find(params[:id])
+    @question = find_question
   end
 
   def create
-    if !current_user
-      flash[:notice] = "You must be signed in to create a new question"
-      render 'new'
-    else
-      title = question_params[:title]
-      description = question_params[:description]
+    user_not_signed_in_notice("create a new question", new_question_path)
 
-      @question = Question.create(user_id: current_user.id, title: title, description: description)
+    if current_user
+      @question = Question.create(question_params)
 
       if @question.save
         flash[:notice] = "You Successfully Created Your New Question!"
@@ -35,15 +35,12 @@ class QuestionsController < ApplicationController
   end
 
   def update
-    @question = Question.find(params[:id])
+    @question = find_question
 
-    if !current_user
-      flash[:notice] = "You must be signed in to edit a question"
-      redirect_to edit_question_path(params[:id])
-    elsif current_user.id != @question.user_id
-      flash[:notice] = "You can only edit your own questions."
-      redirect_to edit_question_path(params[:id])
-    else
+    user_not_signed_in_notice("edit a question", edit_question_path(params[:id]))
+    not_your_question_notice("use edit", edit_question_path(params[:id])) if current_user
+
+    if current_user && current_user.id == @question.user_id
       if @question.update(question_params)
         redirect_to @question
       else
@@ -53,15 +50,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    @question = Question.find(params[:id])
+    @question = find_question
 
-    if !current_user
-      flash[:notice] = "You must be signed in to delete a question."
-      redirect_to @question
-    elsif current_user.id != @question.user_id
-      flash[:notice] = "You can only delete a question you've created."
-      redirect_to @question
-    else
+    user_not_signed_in_notice("delete a question", @question)
+    not_your_question_notice("use delete", @question) if current_user
+
+    if current_user && current_user.id == @question.user_id
       @question.destroy
       answers = Answer.where(question_id: @question.id)
       answers.delete_all
@@ -73,6 +67,8 @@ class QuestionsController < ApplicationController
 
   private
     def question_params
-      params.require(:question).permit(:title, :description)
+      params_hash = params.require(:question).permit(:title, :description)
+      new_hash = { user_id: current_user.id }
+      new_hash.merge!(params_hash)
     end
 end
